requirements.txt
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
requests==2.31.0
pyserial==3.5
esptool==4.7.0 